Agent_Frameworks/LangGraph/requirements.txt
# ── Core LangGraph stack ───────────────────────────────────────────────────────
langgraph>=1.0.5
langgraph-checkpoint-sqlite>=3.0.3
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain>=0.3.0
langchain-experimental>=0.3.0

# ── OpenAI ────────────────────────────────────────────────────────────────────
openai>=1.0.0

# ── Search & scraping tools ───────────────────────────────────────────────────
tavily-python>=0.3.0
duckduckgo-search>=6.0.0
arxiv>=2.1.0
Wikipedia-API>=0.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# ── Configuration & environment ───────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Terminal display ──────────────────────────────────────────────────────────
rich>=13.0.0

# ── Async HTTP ────────────────────────────────────────────────────────────────
aiohttp>=3.9.0
